Move new method to Manager class
see https://github.com/Automattic/jetpack/pull/28208#issuecomment-1377496957
