+++ projects/packages/connection/src/class-manager.php
@@ -2488,4 +2488,22 @@ class Manager {
 		}
 		return $stats;
 	}
+
+	/**
+	 * Get the WPCOM or self-hosted site ID.
+	 *
+	 * @return int|WP_Error
+	 */
+	public static function get_site_id() {
+		$is_wpcom = ( defined( 'IS_WPCOM' ) && IS_WPCOM );
+		$site_id  = $is_wpcom ? get_current_blog_id() : \Jetpack_Options::get_option( 'id' );
+		if ( ! $site_id ) {
+			return new \WP_Error(
+				'unavailable_site_id',
+				__( 'Sorry, something is wrong with your Jetpack connection.', 'jetpack-connection' ),
+				403
+			);
+		}
+		return (int) $site_id;
+	}
 }
